+++ custom_components/larnitech/event.py
@@ -81,7 +81,7 @@ class LarnitechButton(LarnitechEntity, EventEntity):
 
     def __init__(self, coordinator, addr):
         super().__init__(coordinator, addr)
-        self.entity_id = ENTITY_ID_FORMAT.format(self._slug)
+        self.entity_id = ENTITY_ID_FORMAT.format(self._oid())
 
     @callback
     def _handle_coordinator_update(self) -> None: