+++ custom_components/larnitech/valve.py
@@ -57,7 +57,7 @@ class LarnitechValve(LarnitechEntity, ValveEntity):
 
     def __init__(self, coordinator, addr):
         super().__init__(coordinator, addr)
-        self.entity_id = ENTITY_ID_FORMAT.format(self._slug)
+        self.entity_id = ENTITY_ID_FORMAT.format(self._oid())
 
     @property
     def is_closed(self) -> bool: